feat: linearity checking fully implemented
What this means is that all values when assigned to variable either a
binding or function or struct, whatever, have a "must use exactly once"
default. You can annotate the type like: I32@2 by adding the @2 to say
"I actually want to be able to use this twice." Additionally, you can
use I32@? to say "I think I'll use this exactly once, but I might not
use it at all" ... and finally, if you really are a lazy mothertrucker,
you can use I32@* to use that 32-bit integer an infinite number of
times.

If you don't like this, I get it, it's not for everyone, but Orion won't
be the language for you.