feat: cross-compliation... linux, macos, windows
Switched the linker from lld to clang/clang-cl/mingw-w64. This is now
required to have installed for linking. This is for the bootstrap
compiler alone, the full version of Orion will take a different approach
for linking to be more flexible and independent of more vendors.

Now supports a --target <triple> argument to the compiler that will be
used during codegen and linking to build a binary for the target
platform. Might look like this: x86_64-pc-linux-gnu or
aarch64-apple-macos26.0 or something else.